requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
asyncpg==0.29.0
attrs==23.2.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
orjson==3.10.6
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pickleshare==0.7.5
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
pure-eval==0.2.2
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyproject_hooks==1.1.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
wheel==0.43.0
yarg==0.1.9
